+++ example/lib/main.dart
@@ -1,22 +1,23 @@
 import 'package:example/pages/buttons.dart';
 import 'package:example/pages/colors_page.dart';
+import 'package:example/pages/context_menus.dart';
 import 'package:example/pages/dialogs_page.dart';
 import 'package:example/pages/fields.dart';
 import 'package:example/pages/indicators.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:macos_ui/macos_ui.dart';
-import 'package:macos_ui/src/library.dart';
 import 'package:provider/provider.dart';
 
 import 'theme.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider(
@@ -29,7 +30,7 @@ class MyApp extends StatelessWidget {
           darkTheme: MacosThemeData.dark(),
           themeMode: appTheme.mode,
           debugShowCheckedModeBanner: false,
-          home: Demo(),
+          home: const Demo(),
         );
       },
     );
@@ -37,6 +38,8 @@ class MyApp extends StatelessWidget {
 }
 
 class Demo extends StatefulWidget {
+  const Demo({Key? key}) : super(key: key);
+
   @override
   _DemoState createState() => _DemoState();
 }
@@ -50,14 +53,17 @@ class _DemoState extends State<Demo> {
 
   final List<Widget> pages = [
     CupertinoTabView(
-      builder: (_) => ButtonsPage(),
+      builder: (_) => const ButtonsPage(),
     ),
-    IndicatorsPage(),
-    FieldsPage(),
-    ColorsPage(),
-    Center(child: Text('Disclosure item 2')),
-    Center(child: Text('Disclosure item 3')),
-    DialogsPage(),
+    const IndicatorsPage(),
+    const FieldsPage(),
+    const ColorsPage(),
+    const ContextMenusPage(),
+    const Center(
+        child: MacosIcon(
+      CupertinoIcons.add,
+    )),
+    const DialogsPage(),
   ];
 
   Color textLuminance(Color backgroundColor) {
@@ -75,15 +81,12 @@ class _DemoState extends State<Demo> {
       ),
       sidebar: Sidebar(
         minWidth: 200,
-        bottom: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Row(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Icon(CupertinoIcons.profile_circled),
-              const SizedBox(width: 8.0),
-              Text('Tim Apple'),
-            ],
+        bottom: const Padding(
+          padding: EdgeInsets.all(16.0),
+          child: MacosListTile(
+            leading: MacosIcon(CupertinoIcons.profile_circled),
+            title: Text('Tim Apple'),
+            subtitle: Text('tim@apple.com'),
           ),
         ),
         builder: (context, controller) {
@@ -92,38 +95,38 @@ class _DemoState extends State<Demo> {
             onChanged: (i) => setState(() => pageIndex = i),
             scrollController: controller,
             items: [
-              SidebarItem(
-                leading: Icon(CupertinoIcons.square_on_circle),
+              const SidebarItem(
+                leading: MacosIcon(CupertinoIcons.square_on_circle),
                 label: Text('Buttons'),
               ),
-              SidebarItem(
-                leading: Icon(CupertinoIcons.arrow_2_circlepath),
+              const SidebarItem(
+                leading: MacosIcon(CupertinoIcons.arrow_2_circlepath),
                 label: Text('Indicators'),
               ),
-              SidebarItem(
-                leading: Icon(CupertinoIcons.textbox),
+              const SidebarItem(
+                leading: MacosIcon(CupertinoIcons.textbox),
                 label: Text('Fields'),
               ),
-              SidebarItem(
+              const SidebarItem(
                 label: Text('Disclosure'),
                 disclosureItems: [
                   SidebarItem(
-                    leading: Icon(CupertinoIcons.infinite),
+                    leading: MacosIcon(CupertinoIcons.infinite),
                     label: Text('Colors'),
                   ),
                   SidebarItem(
-                    leading: Icon(CupertinoIcons.heart),
-                    label: Text('Item 2'),
+                    leading: MacosIcon(CupertinoIcons.heart),
+                    label: Text('Context Menus'),
                   ),
                   SidebarItem(
-                    leading: Icon(CupertinoIcons.infinite),
+                    leading: MacosIcon(CupertinoIcons.infinite),
                     label: Text('Item 3'),
                   ),
                 ],
               ),
-              SidebarItem(
-                leading: Icon(CupertinoIcons.rectangle),
-                label: Text('Dialogs'),
+              const SidebarItem(
+                leading: MacosIcon(CupertinoIcons.rectangle),
+                label: Text('Dialogs & Sheets'),
               ),
             ],
           );