+++ example/lib/pages/dialogs_page.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:macos_ui/macos_ui.dart';
+import 'package:macos_ui/src/library.dart';
 
 class DialogsPage extends StatefulWidget {
-  DialogsPage({Key? key}) : super(key: key);
+  const DialogsPage({Key? key}) : super(key: key);
 
   @override
   _DialogsPageState createState() => _DialogsPageState();
@@ -12,36 +13,36 @@ class _DialogsPageState extends State<DialogsPage> {
   @override
   Widget build(BuildContext context) {
     return MacosScaffold(
-      titleBar: TitleBar(
-        title: Text('macOS UI Dialogs'),
+      titleBar: const TitleBar(
+        title: const Text('macOS UI Dialogs & Sheets'),
       ),
       children: [
         ContentArea(builder: (context, scrollController) {
           return SingleChildScrollView(
             controller: scrollController,
-            padding: EdgeInsets.all(20),
+            padding: const EdgeInsets.all(20),
             child: Center(
               child: Column(
                 children: [
                   PushButton(
                     buttonSize: ButtonSize.large,
-                    child: Text('Show Alert Dialog 1'),
+                    child: const Text('Show Alert Dialog 1'),
                     onPressed: () => showMacosAlertDialog(
                       context: context,
                       builder: (context) => MacosAlertDialog(
-                        appIcon: FlutterLogo(
+                        appIcon: const FlutterLogo(
                           size: 56,
                         ),
-                        title: Text(
+                        title: const Text(
                           'Alert Dialog with Primary Action',
                         ),
-                        message: Text(
+                        message: const Text(
                           'This is an alert dialog with a primary action and no secondary action',
                         ),
                         //horizontalActions: false,
                         primaryButton: PushButton(
                           buttonSize: ButtonSize.large,
-                          child: Text('Primary'),
+                          child: const Text('Primary'),
                           onPressed: Navigator.of(context).pop,
                         ),
                       ),
@@ -50,39 +51,30 @@ class _DialogsPageState extends State<DialogsPage> {
                   const SizedBox(height: 16),
                   PushButton(
                     buttonSize: ButtonSize.large,
-                    child: Text('Show Alert Dialog 2'),
+                    child: const Text('Show Alert Dialog 2'),
                     onPressed: () => showMacosAlertDialog(
                       context: context,
                       builder: (context) => MacosAlertDialog(
-                        appIcon: FlutterLogo(
+                        appIcon: const FlutterLogo(
                           size: 56,
                         ),
-                        title: Text(
+                        title: const Text(
                           'Alert Dialog with Secondary Action',
                         ),
-                        message: Text(
+                        message: const Text(
                           'This is an alert dialog with primary action and secondary action laid out horizontally',
                           textAlign: TextAlign.center,
                         ),
                         //horizontalActions: false,
                         primaryButton: PushButton(
                           buttonSize: ButtonSize.large,
-                          child: Text('Primary'),
+                          child: const Text('Primary'),
                           onPressed: Navigator.of(context).pop,
                         ),
                         secondaryButton: PushButton(
                           buttonSize: ButtonSize.large,
-                          color: MacosTheme.brightnessOf(context).isDark
-                              ? MacosColors.controlColor.darkColor
-                              : MacosColors.controlColor,
-                          child: Text(
-                            'Secondary',
-                            style: TextStyle(
-                              color: MacosTheme.brightnessOf(context).isDark
-                                  ? MacosColors.controlTextColor.darkColor
-                                  : MacosColors.controlTextColor,
-                            ),
-                          ),
+                          isSecondary: true,
+                          child: const Text('Secondary'),
                           onPressed: Navigator.of(context).pop,
                         ),
                       ),
@@ -91,39 +83,30 @@ class _DialogsPageState extends State<DialogsPage> {
                   const SizedBox(height: 16),
                   PushButton(
                     buttonSize: ButtonSize.large,
-                    child: Text('Show Alert Dialog 3'),
+                    child: const Text('Show Alert Dialog 3'),
                     onPressed: () => showMacosAlertDialog(
                       context: context,
                       builder: (context) => MacosAlertDialog(
-                        appIcon: FlutterLogo(
+                        appIcon: const FlutterLogo(
                           size: 56,
                         ),
-                        title: Text(
+                        title: const Text(
                           'Alert Dialog with Secondary Action',
                         ),
-                        message: Text(
+                        message: const Text(
                           'This is an alert dialog with primary action and secondary action laid out vertically',
                           textAlign: TextAlign.center,
                         ),
                         horizontalActions: false,
                         primaryButton: PushButton(
                           buttonSize: ButtonSize.large,
-                          child: Text('Primary'),
+                          child: const Text('Primary'),
                           onPressed: Navigator.of(context).pop,
                         ),
                         secondaryButton: PushButton(
                           buttonSize: ButtonSize.large,
-                          color: MacosTheme.brightnessOf(context).isDark
-                              ? MacosColors.controlColor.darkColor
-                              : MacosColors.controlColor,
-                          child: Text(
-                            'Secondary',
-                            style: TextStyle(
-                              color: MacosTheme.brightnessOf(context).isDark
-                                  ? MacosColors.controlTextColor.darkColor
-                                  : MacosColors.controlTextColor,
-                            ),
-                          ),
+                          isSecondary: true,
+                          child: const Text('Secondary'),
                           onPressed: Navigator.of(context).pop,
                         ),
                       ),
@@ -132,17 +115,17 @@ class _DialogsPageState extends State<DialogsPage> {
                   const SizedBox(height: 16),
                   PushButton(
                     buttonSize: ButtonSize.large,
-                    child: Text('Show Alert Dialog 4'),
+                    child: const Text('Show Alert Dialog 4'),
                     onPressed: () => showMacosAlertDialog(
                       context: context,
                       builder: (context) => MacosAlertDialog(
-                        appIcon: FlutterLogo(
+                        appIcon: const FlutterLogo(
                           size: 56,
                         ),
-                        title: Text(
+                        title: const Text(
                           'Alert Dialog with Secondary Action',
                         ),
-                        message: Text(
+                        message: const Text(
                           'This is an alert dialog with primary action and secondary '
                           'action laid out vertically. It also contains a "suppress" option.',
                           textAlign: TextAlign.center,
@@ -150,29 +133,29 @@ class _DialogsPageState extends State<DialogsPage> {
                         horizontalActions: false,
                         primaryButton: PushButton(
                           buttonSize: ButtonSize.large,
-                          child: Text('Primary'),
+                          child: const Text('Primary'),
                           onPressed: Navigator.of(context).pop,
                         ),
                         secondaryButton: PushButton(
                           buttonSize: ButtonSize.large,
-                          color: MacosDynamicColor.resolve(
-                            MacosColors.controlColor,
-                            context,
-                          ),
-                          child: Text(
-                            'Secondary',
-                            style: TextStyle(
-                              color: MacosDynamicColor.resolve(
-                                MacosColors.controlColor,
-                                context,
-                              ),
-                            ),
-                          ),
+                          isSecondary: true,
+                          child: const Text('Secondary'),
                           onPressed: Navigator.of(context).pop,
                         ),
-                        suppress: DoNotNotifyRow(),
+                        suppress: const DoNotNotifyRow(),
                       ),
                     ),
+                  ),
+                  const SizedBox(height: 16),
+                  PushButton(
+                    buttonSize: ButtonSize.large,
+                    child: const Text('Show sheet'),
+                    onPressed: () {
+                      showMacosSheet(
+                        context: context,
+                        builder: (_) => const MacosuiSheet(),
+                      );
+                    },
                   ),
                 ],
               ),
@@ -206,8 +189,72 @@ class _DoNotNotifyRowState extends State<DoNotNotifyRow> {
           },
         ),
         const SizedBox(width: 8),
-        Text('Don\'t ask again'),
+        const Text('Don\'t ask again'),
       ],
+    );
+  }
+}
+
+class MacosuiSheet extends StatelessWidget {
+  const MacosuiSheet({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return MacosSheet(
+      child: Center(
+        child: Column(
+          children: [
+            const SizedBox(height: 50),
+            const FlutterLogo(
+              size: 56,
+            ),
+            const SizedBox(height: 24),
+            Text(
+              'Welcome to macos_ui',
+              style: MacosTheme.of(context).typography.largeTitle.copyWith(
+                    fontWeight: FontWeight.bold,
+                  ),
+            ),
+            const SizedBox(height: 24),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                const MacosListTile(
+                  leading: MacosIcon(CupertinoIcons.lightbulb),
+                  title: Text(
+                    'A robust library of Flutter components for macOS',
+                    //style: MacosTheme.of(context).typography.headline,
+                  ),
+                  subtitle: Text(
+                    'Create native looking macOS applications using Flutter',
+                  ),
+                ),
+              ],
+            ),
+            const SizedBox(height: 16),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                const MacosListTile(
+                  leading: MacosIcon(CupertinoIcons.bolt),
+                  title: Text(
+                    'Create beautiful macOS applications in minutes',
+                    //style: MacosTheme.of(context).typography.headline,
+                  ),
+                ),
+                const SizedBox(width: 10),
+              ],
+            ),
+            const Spacer(),
+            PushButton(
+              buttonSize: ButtonSize.large,
+              child: const Text('Get started'),
+              onPressed: () => Navigator.of(context).pop(),
+            ),
+            const SizedBox(height: 50),
+          ],
+        ),
+      ),
     );
   }
 }