+++ lib/src/fields/text_field.dart
@@ -116,7 +116,9 @@ class _TextFieldSelectionGestureDetectorBuilder
         return;
       }
     }
-    super.onSingleTapUp(details);
+    if (delegate.selectionEnabled) {
+      renderEditable.selectPosition(cause: SelectionChangedCause.tap);
+    }
     _state._requestKeyboard();
     if (_state.widget.onTap != null) _state.widget.onTap!();
   }
@@ -1302,7 +1304,7 @@ class _MacosTextFieldState extends State<MacosTextField>
         }
         return focusedDecoration.border;
       }(),
-      color: focusedDecoration.color ?? Color(0x00000000),
+      color: focusedDecoration.color ?? const Color(0x00000000),
     );
 
     final Color selectionColor =