+++ lib/src/library.dart
@@ -20,6 +20,9 @@ export 'package:flutter/material.dart'
         Divider,
         SelectableText,
         VisualDensity,
+        Colors,
+        MaterialLocalizations,
+        Dialog,
         kElevationToShadow;
 export 'package:flutter/widgets.dart';
 export 'util.dart';